requirements.txt
absl-py==0.7.1
numpy   astor==0.7.1
atomicwrites==1.3.0
attrs==19.1.0
backcall==0.1.0
backports-abc==0.5
blosc==1.8.1
boto3==1.9.149
botocore==1.12.149
cachetools==3.1.0
certifi==2019.3.9
chardet==3.0.4
Click==7.0
cloud-volume==0.49.0
compressed-segmentation==1.0.0
connected-components-3d==1.0.7
coverage==4.5.3
cycler==0.10.0
Cython==0.29.7
decorator==4.4.0
dijkstra3d==1.0.1
dill==0.2.9
docutils==0.14
DracoPy==0.0.10
-e git+https://github.com/data-exchange/dxchange#egg=dxchange
edt==1.3.2
fasteners==0.14.1
fastremap==1.5.2
ffn==0.1.0
fpzip==1.1.3
gast==0.2.2
gevent==1.4.0
google-api-core==1.11.0
google-apitools==0.5.28
google-auth==1.6.3
google-cloud-core==1.0.0
google-cloud-logging==1.10.1
google-cloud-storage==1.15.1
google-resumable-media==0.3.2
googleapis-common-protos==1.5.10
greenlet==0.4.15
grpcio==1.20.1
h5py==2.9.0
httplib2==0.12.3
idna==2.8
igneous==0.0.1.dev237
imageio==2.5.0
inflection==0.3.1
intern==0.9.11
ipykernel==5.1.1
ipython==7.5.0
ipython-genutils==0.2.0
jedi==0.13.3
jmespath==0.9.4
json5==0.5.1
jsonschema==2.6.0
jupyter-client==5.2.4
jupyter-core==4.4.0
Keras-Applications==1.0.7
Keras-Preprocessing==1.0.9
kimimaro==0.4.1
kiwisolver==1.1.0
knossos-utils==0.1
Markdown==2.6.11
matplotlib==3.0.3
mock==3.0.5
monotonic==1.5
more-itertools==7.0.0
mpi4py==3.0.1
multiprocess==0.70.7
networkx==2.3
neuroglancer==1.0.15
nose2==0.9.1
numpy==1.16.3
oauth2client==3.0.0
opencv-python==4.1.0.25
pandas==0.24.2
parso==0.4.0
pathos==0.2.3
pbr==5.2.0
pexpect==4.7.0
pickleshare==0.7.5
Pillow==6.0.0
Pillow-PIL==0.1.dev0
pluggy==0.11.0
posix-ipc==1.0.4
pox==0.2.5
ppft==1.6.4.9
prompt-toolkit==2.0.9
protobuf==3.7.1
psutil==5.4.3
ptyprocess==0.6.0
py==1.8.0
pyasn1==0.4.5
pyasn1-modules==0.2.5
Pygments==2.4.0
pyparsing==2.4.0
pytest==4.5.0
python-dateutil==2.8.0
python-jsonschema-objects==0.3.5
python-snappy==0.5.4
pytz==2019.1
PyWavelets==1.0.3
pyzmq==18.0.1
requests==2.22.0
rsa==4.0
s3transfer==0.2.0
scikit-image==0.15.0
scipy==1.2.1
six==1.12.0
sockjs-tornado==1.0.6
task-queue==0.14.4
tenacity==5.0.4
tensorboard==1.13.1
tensorflow==1.13.1
tensorflow-estimator==1.13.0
termcolor==1.1.0
tifffile==2019.3.18
tinybrain==0.1.0
tornado==5.1.1
tqdm==4.32.1
traitlets==4.3.2
urllib3==1.24.2
wcwidth==0.1.7
Werkzeug==0.15.4
zmesh==0.2.0
